+++ src/benchmarking/counters.h
@@ -21,11 +21,11 @@ using namespace Diperan::Benchmark::Results;
 namespace Diperan {
     namespace Benchmarks {
         void prepare_counters(std::vector<int32_t>& event_codes);
-        void read_counters(const std::vector<int32_t>& event_codes, std::vector<int64_t>& counter_values);
+        void read_counters(const std::vector<int32_t>& event_codes, std::vector<long long int>& counter_values);
         void existing_events(std::vector<int32_t>& event_codes);
         bool check_event_exists(int32_t event_code);
         void create_benchmark_results_flatb(const std::vector <int32_t> &event_codes,
-                                            const std::vector <int64_t> &counter_values,
+                                            const std::vector <long long int> &counter_values,
                                             int64_t cycles,
                                             int64_t time,
                                             std::string benchmark_type);
